+++ fiberplane-models/src/lib.rs
@@ -12,6 +12,7 @@ products, including but not limited to:
 - Providers
   - Provider Schemas
 - Comments
+- Front Matter Schemas
 - Rich-Text Formatting
 - Templates
 - Views
@@ -19,7 +20,7 @@ products, including but not limited to:
 
 */
 
-use serde::de::{Error, Unexpected, Visitor};
+use serde::de::{self, Error, Unexpected, Visitor};
 use serde::{Deserialize, Deserializer};
 use std::fmt;
 
@@ -29,6 +30,7 @@ pub mod data_sources;
 pub mod events;
 pub mod files;
 pub mod formatting;
+pub mod front_matter_schemas;
 pub mod integrations;
 pub mod labels;
 pub mod names;
@@ -115,3 +117,16 @@ impl<'de> Visitor<'de> for U32Visitor {
             .map_err(|_| Error::invalid_type(Unexpected::Str(&v), &self))
     }
 }
+
+/// Helper trait to use as a trait bound on generic types that must be serializable when
+/// the `fp-bindgen` feature is enabled
+pub trait BindgenSerializable {}
+
+#[cfg(feature = "fp-bindgen")]
+impl<T> BindgenSerializable for T where
+    T: fp_bindgen::prelude::Serializable + for<'de> de::Deserialize<'de>
+{
+}
+
+#[cfg(not(feature = "fp-bindgen"))]
+impl<T> BindgenSerializable for T {}